feat: autosave editor, and warn when leaving without saving
